when in editing mode, if image is changed, thumbnail is deleted so check for thumbnail existance when updating thumbnail so we don't get php error for trying it twice

+++ src/UNL/ENews/Controller.php
@@ -213,9 +213,10 @@ class UNL_ENews_Controller
                     header('Location: ?view=thanks&_type=story');
                     exit();
                 }
-                //Delete exisiting thumbnail
-                $thumb = $story->getFileByUse('thumbnail');
-                $thumb->delete();
+                //Delete existing thumbnail
+                if ($thumb = $story->getFileByUse('thumbnail')) {
+                    $thumb->delete();
+                }
                 $mysqli = UNL_ENews_Controller::getDB();
                 $sql = 'DELETE FROM story_files WHERE story_id = '.intval($story->id).' AND file_id = '.intval($thumb->id);
                 $mysqli->query($sql);
